requirements.txt
Pillow==10.3.0
faiss-cpu==1.12.0
sentence-transformers==2.6.1
groq==0.9.0
openai==1.35.0
streamlit==1.35.0
pypdf2==3.0.1
python-dotenv==1.0.1
numpy==1.26.4
